chore: renaming project to 'algobase'
