+++ .github/workflows/publish-to-pypi.yml
@@ -63,7 +63,7 @@ jobs:
     runs-on: ubuntu-latest
     environment:
       name: pypi
-      url: https://pypi.org/p/humblepy
+      url: https://pypi.org/p/algobase
     permissions:
       id-token: write # IMPORTANT: mandatory for trusted publishing
