+++ algobase/types/__init__.py
@@ -1,0 +1,1 @@
+"""Types for algobase."""
