+++ algobase/utils/hash.py
