--- humblepy/types/__init__.py
@@ -1,1 +1,0 @@
-"""Types for humblepy."""
