+++ pyproject.toml
@@ -4,14 +4,14 @@ requires = ["poetry_core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
-name = "humblepy"
-version = "0.6.2"
+name = "algobase"
+version = "0.7.0"
 description = "A type-safe Python library for interacting with assets on Algorand."
 readme = "README.md"
-authors = ["humblepy <alexandercodes@proton.me>"]
+authors = ["algobase <alexandercodes@proton.me>"]
 license = "Apache Software License 2.0"
-repository = "https://github.com/code-alexander/humblepy"
-homepage = "https://github.com/code-alexander/humblepy"
+repository = "https://github.com/code-alexander/algobase"
+homepage = "https://github.com/code-alexander/algobase"
 
 # Keywords description https://python-poetry.org/docs/pyproject/#keywords
 keywords = []  #! Update me
@@ -29,7 +29,7 @@ classifiers = [  #! Update me
 
 [tool.poetry.scripts]
 # Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
-# "humblepy" = "humblepy.__main__:app"
+# "algobase" = "algobase.__main__:app"
 
 [tool.poetry.dependencies]
 python = "^3.11"
@@ -122,7 +122,7 @@ addopts = [
 source = ["tests"]
 
 [coverage.paths]
-source = "humblepy"
+source = "algobase"
 
 [coverage.run]
 branch = true
