+++ requirements.txt
@@ -3,7 +3,7 @@ babel==2.14.0 ; python_version >= "3.11" and python_version < "4.0"
 cffi==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
 click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
 colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0"
-cytoolz==0.12.2 ; python_version >= "3.11" and python_version < "4.0"
+cytoolz==0.12.3 ; python_version >= "3.11" and python_version < "4.0"
 markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
 mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
 msgpack==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
@@ -16,6 +16,6 @@ pygments==2.17.2 ; python_version >= "3.11" and python_version < "4.0"
 pynacl==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
 rich==13.7.0 ; python_version >= "3.11" and python_version < "4.0"
 shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
-toolz==0.12.0 ; python_version >= "3.11" and python_version < "4.0"
+toolz==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
 typer[all]==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
 typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
