+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Tests for humblepy."""
+"""Tests for algobase."""
